src/app_baseball_agent/pyproject.toml
[tool.poetry]
name = "app-baseball-agent"
version = "0.1.0"
description = ""
authors = ["Simon Jakesch <simonjakesch@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.32.0"
asyncio = "^3.4.3"
asyncpg = "^0.30.0"
llama-index = "^0.11.23"
llama-index-llms-azure-openai = "^0.2.2"
llama-index-embeddings-azure-openai = "^0.2.5"
psycopg2-binary = "^2.9.10"
fastapi = "^0.115.5"
python-dotenv = "^1.0.1"
llama-index-tools-azure-code-interpreter = "^0.3.0"
az-cli = "^0.5"
ollama = "^0.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
